Converted to parameter splatting between scripts
Parameter splatting is a method of passing multiple parameters to a
command with one variable. As implemented, the method also makes it
possible so that parameters can be passed between scripts without
needing to know what parameters are required by the downstream script.
This is very handy is an extensible bootstrapping framework such as this
one.


Former-commit-id: f607c8f5b7a710b85339c33d844dab25afe31d87