Added functions and created linuxsaltinstall shell
Added functions, createWorkingDir() and downloadFile().
createWorkingDir() doesn't seem to be working. Getting an error that the
`workingdir` variable isn't globally defined, but it isn't supposed to
be at that point.

Added some logic to loop through the scriptstoexecute, download them,
and execute them.

Created a shell for the LinuxSaltInstall script