Fix line wraps in log file
Start-Transcript wraps lines at the width of the window, which defaults
to 120 characters. This is particularly difficult to modify in a
non-interactive script (which has no real window). We accomplish this by
using functions to modify the registry setting for the powershell.exe
console that sets the width of the window, and then to modify the width
of the current powershell session.