Fix line wraps in log file
Start-Transcript wraps lines at the width of the window, which defaults
to 120 characters. This is particularly difficult to modify in a
non-interactive script (which has no real window). We accomplish this by
using functions to modify the registry setting for the powershell.exe
console that sets the width of the window, and then to modify the width
of the current powershell session.


Former-commit-id: c6d5cba174c24894434c32b503b5eb0a1ea9517b
Former-commit-id: 6216d84eb6557f05f6c0a1f40c33c7d7f52bb1aa
Former-commit-id: 877b512a43c313f5336d8c41c967bcaa8e90fa95