subprocess can now run fullcommand as is
However, it is not usually a good idea to use 'shell=True' but since we
build the command to be processed, it can be okay.  However, I will
continue to explore options ... the biggest issue with a list of args
seems to be with how yumrepomap is passed in ... it appears I may have
found a potential issue with the yumrepo script ... will explore.