Moving argparse out of the class itself
Adam Fest pointed out that this was kind of silly and I agree with him
as argparse would only be viable in the case of running the script from
the CLI. Updated yumrepoinstall to use json.loads() but must be sure to
pass in youmrepomap correctly! Also updated subprocess call on Linux
side to use list of args so we don't need shell=True, but still need to
work on Windows side.