Uses default aws path for aws binary
* Aligns the path for the aws binary with Amazon Linux
* Tests for the aws binary and installs the awscli only if the
binary does not exist and is not executable
* Exports an env for the AWS_DEFAULT_REGION

This patch is intended to eliminate circumstances where an extra
instance of the awscli gets installed. Previously, the awscli would
be installed regardless of the prior state, possibly resulting in
multiple instances of the awscli and causing confusion about which
version of the awscli should be used.